Status effect system

+++ Assets/Scripts/Health/HealthEvent.cs
@@ -8,7 +8,7 @@ public class HealthEvent : MonoBehaviour
 {
     public event Action<HealthEvent, HealthArg> OnHealth;
 
-    public void CallHealthEvent(float percentHealth, int currentHealth, int maxHealth,int defence, bool isDead)
+    public void CallHealthEvent(float percentHealth, float currentHealth, int maxHealth,int defence, bool isDead)
     {
         OnHealth?.Invoke(this, new HealthArg()
         {
@@ -23,7 +23,7 @@ public class HealthEvent : MonoBehaviour
 public class HealthArg: EventArgs
 {
     public float percentHealth;
-    public int currentHealth;
+    public float currentHealth;
     public int maxHealth;
     public int defence;
     public bool isDead;
